+++ src/utils/Helpers.js
@@ -1,7 +1,9 @@
 var DEBUG_ENABLED = require('../config').DEBUG_ENABLED,
-  DATA_TYPES = require('../constants/DataTypes'),
-  ErrorMessages = require('../constants/ErrorMessages'),
-  CAS = require('../constants/CASConstants');
+		DATA_TYPES = require('../constants/DataTypes'),
+		ErrorMessages = require('../constants/ErrorMessages'),
+		CAS = require('../constants/CASConstants'),
+// Default delimiter to wrap character strings in SQL queries.
+		DEFAULT_DELIMITER = "'";
 
 /**
  * Emit event only if there are registered listeners for the event
@@ -200,40 +202,41 @@ exports._escapeString = _escapeString;
  * @private
  */
 exports._sqlFormat = function (sql, arrValues, arrDelimiters) {
-  arrValues = [].concat(arrValues);
+	if (!Array.isArray(arrValues)) {
+    arrValues = [arrValues];
+	}
 
-	arrDelimiters = arrDelimiters || [];
+	if (!Array.isArray(arrDelimiters)) {
+		arrDelimiters = [arrDelimiters];
+	}
 
-  if (arrDelimiters.length) {
-    arrDelimiters = [].concat(arrDelimiters);
-  } else {
-	  // By default wrap all values with a single quote.
-	  // CUBRID Server will automatically cast values depending
-	  // on the column data type.
-    for (var i = arrValues.length; i > 0; --i) {
-      arrDelimiters.push("'");
-    }
-  }
+	var i = -1,
+			valCount = arrValues.length,
+			delimitersCount = arrDelimiters.length;
 
-  return sql.replace(/\?/g, function (match) {
-    if (!arrValues.length) {
+	return sql.replace(/\?/g, function (match) {
+    if (++i == valCount) {
       return match;
     }
 
-	  // Get the front element.
-    var val = arrValues.shift(),
-    // And the front delimiter.
-      delimiter = arrDelimiters.shift();
+		// Get the value for the current placeholder.
+		// We iterate via `i` instead of shifting from the front of the
+		// array because we do not want to alter the original array
+		// received from the application. The application may
+		// choose to reuse it in the loop.
+		var val = arrValues[i],
+		// And its delimiter. If not defined, use single quotes.
+				delimiter = (i >= delimitersCount ? DEFAULT_DELIMITER : arrDelimiters[i]);
 
-    if (val === undefined || val === null) {
-      return 'NULL';
-    }
+		if (val === undefined || val === null) {
+		  return 'NULL';
+		}
 
-	  // Send numbers as real numbers. Numbers wrapped in strings
-	  // are not considered as numbers. They are sent as strings.
-    if (typeof val === 'number') {
-      return val;
-    }
+		// Send numbers as real numbers. Numbers wrapped in strings
+		// are not considered as numbers. They are sent as strings.
+		if (typeof val === 'number') {
+		  return val;
+		}
 
 	  // If the value is of Date type, convert it into
 	  // CUBRID compatible DATETIME format strings.
@@ -245,14 +248,19 @@ exports._sqlFormat = function (sql, arrValues, arrDelimiters) {
 		  // Broker we choose the
 		  // `'mm/dd[/yyyy] hh:mi[:ss[.ff]] [am|pm]'` format.
 
-      return "'" +
+      return DEFAULT_DELIMITER +
 			    // Month value in JavaScript is 0 based, i.e. 0-11,
 			    // but CUBRID is 1-12. Also CUBRID doesn't care if
 		      // dates are zero-padded or not.
 		      (val.getMonth() + 1) + '/' + val.getDate() + '/' + val.getFullYear() +
 		      ' ' + val.getHours() + ':' + val.getMinutes() + ':' + val.getSeconds() +
-		      '.' + val.getMilliseconds() + "'";
+		      '.' + val.getMilliseconds() + DEFAULT_DELIMITER;
     }
+
+		// Delimiters must be specified as strings.
+		if (typeof delimiter !== 'string') {
+			delimiter = DEFAULT_DELIMITER;
+		}
 
 	  // Otherwise, safely escape the string.
     return delimiter + _escapeString(val) + delimiter;